app/requirements.txt
# requirements.txt
Flask==2.1.1
mysql-connector-python==8.0.27
